requirements.txt
absl-py==1.0.0
asn1crypto==1.0.1
astor==0.8.1
beautifulsoup4==4.8.0
bleach==1.5.0
bs4==0.0.1
cffi==1.12.3
cryptography==2.7
cycler==0.10.0
gast==0.4.0
grpcio==1.42.0
h5py==2.9.0
html5lib==0.9999999
idna==2.8
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.1.0
Markdown==3.3.6
matplotlib==3.1.1
mock==3.0.5
nltk==3.4.5
numpy==1.17.2
overrides==3.1.0
pandas==0.25.1
pbr==4.2.0
Pillow==6.2.0
pip==19.2.3
protobuf==3.19.1
pycparser==2.19
pyparsing==2.4.2
python-dateutil==2.8.0
pytz==2019.3
regex==2018.6.6
scipy==1.3.1
setuptools==41.4.0
six==1.12.0
stevedore==1.29.0
tensorboard==1.13.1
tensorflow-estimator==1.13.0
tensorflow-tensorboard==1.5.1
termcolor==1.1.0
torch==1.7.0
torchvision==0.8.1
transitions==0.8.6
typing-utils==0.0.3
virtualenv==16.0.0
virtualenv-clone==0.3.0
virtualenvwrapper==4.8.2
Werkzeug==0.16.0
wheel==0.33.6
